+++ packages/xchain-doge/src/clientLedger.ts
@@ -1,0 +1,95 @@
+import AppBtc from '@ledgerhq/hw-app-btc'
+import { Transaction } from '@ledgerhq/hw-app-btc/lib/types'
+import { FeeOption, FeeRate, TxHash, TxParams } from '@xchainjs/xchain-client'
+import { Address } from '@xchainjs/xchain-util'
+import { UtxoClientParams } from '@xchainjs/xchain-utxo'
+
+import { Client } from './client'
+/**
+ * Custom Ledger Bitcoin client
+ */
+class ClientLedger extends Client {
+  // Reference to the Ledger transport object
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  private transport: any // TODO: Parametrize
+  private app: AppBtc | undefined
+
+  // Constructor
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  constructor(params: UtxoClientParams & { transport: any }) {
+    super(params)
+    this.transport = params.transport
+  }
+
+  // Get the Ledger Doge application instance
+  public async getApp(): Promise<AppBtc> {
+    if (this.app) {
+      return this.app
+    }
+    this.app = new AppBtc({ transport: this.transport, currency: 'dogecoin' })
+    return this.app
+  }
+
+  // Get the current address synchronously
+  getAddress(): string {
+    throw Error('Sync method not supported for Ledger')
+  }
+
+  // Get the current address asynchronously
+  async getAddressAsync(index = 0, verify = false): Promise<Address> {
+    const app = await this.getApp()
+    const result = await app.getWalletPublicKey(this.getFullDerivationPath(index), {
+      format: 'legacy',
+      verify,
+    })
+    return result.bitcoinAddress
+  }
+
+  // Transfer Doge from Ledger
+  async transfer(params: TxParams & { feeRate?: FeeRate }): Promise<TxHash> {
+    const app = await this.getApp()
+    const fromAddressIndex = params?.walletIndex || 0
+    const derivePath = this.getFullDerivationPath(fromAddressIndex)
+    // Get fee rate
+    const feeRate = params.feeRate || (await this.getFeeRates())[FeeOption.Fast]
+    // Get sender address
+    const sender = await this.getAddressAsync(fromAddressIndex)
+    // Prepare transaction
+    const { psbt, utxos } = await this.buildTx({
+      ...params,
+      feeRate,
+      sender,
+    })
+    // Prepare Ledger inputs
+    const inputs: Array<[Transaction, number, string | null, number | null]> = utxos.map(({ txHex, hash, index }) => {
+      if (!txHex) {
+        throw Error(`Missing 'txHex' for UTXO (txHash ${hash})`)
+      }
+      const splittedTx = app.splitTransaction(txHex, false /* no segwit support */)
+      return [splittedTx, index, null, null]
+    })
+
+    const associatedKeysets: string[] = inputs.map((_) => derivePath)
+    const newTxHex = psbt.data.globalMap.unsignedTx.toBuffer().toString('hex')
+    const newTx: Transaction = app.splitTransaction(newTxHex, true)
+
+    const outputScriptHex = app.serializeTransactionOutputs(newTx).toString('hex')
+
+    const txHex = await app.createPaymentTransaction({
+      inputs,
+      associatedKeysets,
+      outputScriptHex,
+      // no additionals - similar to https://github.com/shapeshift/hdwallet/blob/a61234eb83081a4de54750b8965b873b15803a03/packages/hdwallet-ledger/src/bitcoin.ts#L222
+      additionals: [],
+    })
+
+    const txHash = await this.broadcastTx(txHex)
+    // Throw error if no transaction hash is received
+    if (!txHash) {
+      throw Error('No Tx hash')
+    }
+
+    return txHash
+  }
+}
+export { ClientLedger }