+++ packages/xchain-thorchain-amm/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [8e82dc3]
+- Updated dependencies [8e82dc3]
+- Updated dependencies [2f7988a]
+  - @xchainjs/xchain-doge@0.7.13
+  - @xchainjs/xchain-bitcoincash@0.17.11
+  - @xchainjs/xchain-litecoin@0.14.0
+
 ## 1.1.2
 
 ### Patch Changes